Fix case of false deletion
Only cleanup an old title object if it really gets unrelated with the custom update.

+++ projects/bp-strapi/src/api/picture/content-types/picture/lifecycles.js
@@ -47,7 +47,8 @@ module.exports = {
       where: { title: oldTitleInDB.id },
     });
 
-    if (picturesWithOldTitle.length < 2) {
+    // Cleanup old title if it is going to be unrelated after our custom update
+    if (titleIdForUpdate !== oldTitleInDB.id && picturesWithOldTitle.length < 2) {
       await titleQuery.delete({
         where: { id: oldTitleInDB.id },
       });
